added notification system

+++ database/migrations/2022_07_26_093605_create_comments_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->text('content')->nullable();
             $table->string('comment_image')->nullable();
             $table->string('comment_image_slug')->nullable();
-            $table->integer('votes')->default(0);
+            $table->smallInteger('votes')->default(0);
             $table->timestamps();
         });
     }
